+++ src/components/MobileMenu.tsx
@@ -20,6 +20,7 @@ import { LibraryButton } from "./LibraryButton";
 import { PenModeButton } from "./PenModeButton";
 import { Stats } from "./Stats";
 import { actionToggleStats } from "../actions";
+import { SidebarHeaderComponents } from "./Sidebar/SidebarHeader";
 
 type MobileMenuProps = {
   appState: AppState;
